+++ free-ton/contracts/staking/interfaces/IRelayRound.sol
@@ -10,14 +10,14 @@ interface IRelayRound {
 
     struct RelayRoundDetails {
         address root;
-        uint128 round_num;
+        uint32 round_num;
         Relay[] relays;
         bool relays_installed;
         uint32 code_version;
     }
 
     function getDetails() external view responsible returns (RelayRoundDetails);
-    function sendRelaysToRelayRound(address relay_round_addr, uint128 relays_count, address send_gas_to) external;
+    function sendRelaysToRelayRound(address relay_round_addr, uint32 relays_count, address send_gas_to) external;
     function setRelays(Relay[] _relay_list, address send_gas_to) external;
     function setEmptyRelays(address send_gas_to) external;
     function getRewardForRound(address staker_addr, address send_gas_to, uint32 code_version) external;
